+++ plugins/resources/src/components/panel/details/actions/Logs.tsx
@@ -11,13 +11,13 @@ import {
   TextInput,
 } from '@patternfly/react-core';
 import React, { useContext, useState } from 'react';
-import { IRow } from '@patternfly/react-table';
 import { V1Pod } from '@kubernetes/client-node';
 import { Terminal as xTerm } from 'xterm';
 
 import {
   IPluginsContext,
   IResource,
+  IResourceRow,
   ITerminalContext,
   PluginsContext,
   TERMINAL_OPTIONS,
@@ -52,7 +52,7 @@ const getContainers = (pod: V1Pod): string[] => {
 
 interface ILogsProps {
   request: IResource;
-  resource: IRow;
+  resource: IResourceRow;
   show: boolean;
   setShow: (value: boolean) => void;
 }
@@ -83,9 +83,9 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
       const host = configuredWebSocketAddress || `wss://${window.location.host}`;
 
       const ws = new WebSocket(
-        `${host}/api/plugins/resources/logs?cluster=${resource.cluster.title}${
-          resource.namespace ? `&namespace=${resource.namespace.title}` : ''
-        }&name=${resource.name.title}&container=${container}&since=${since}&tail=${
+        `${host}/api/plugins/resources/logs?cluster=${resource.cluster}${
+          resource.namespace ? `&namespace=${resource.namespace}` : ''
+        }&name=${resource.name}&container=${container}&since=${since}&tail=${
           TERMINAL_OPTIONS.scrollback
         }&previous=false&follow=true`,
       );
@@ -105,7 +105,7 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
       };
 
       terminalsContext.addTerminal({
-        name: `${resource.name.title}: ${container}`,
+        name: `${resource.name}: ${container}`,
         terminal: term,
         webSocket: ws,
       });
@@ -113,7 +113,7 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
       if (err.message) {
         term.write(`${err.message}\n\r`);
         terminalsContext.addTerminal({
-          name: `${resource.name.title}: ${container}`,
+          name: `${resource.name}: ${container}`,
           terminal: term,
         });
       }
@@ -126,9 +126,9 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
 
     try {
       const response = await fetch(
-        `/api/plugins/resources/logs?cluster=${resource.cluster.title}${
-          resource.namespace ? `&namespace=${resource.namespace.title}` : ''
-        }&name=${resource.name.title}&container=${container}&regex=${encodeURIComponent(regex)}&since=${since}&tail=${
+        `/api/plugins/resources/logs?cluster=${resource.cluster}${
+          resource.namespace ? `&namespace=${resource.namespace}` : ''
+        }&name=${resource.name}&container=${container}&regex=${encodeURIComponent(regex)}&since=${since}&tail=${
           TERMINAL_OPTIONS.scrollback
         }&previous=${previous}&follow=false`,
         { method: 'get' },
@@ -138,7 +138,7 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
       if (response.status >= 200 && response.status < 300) {
         term.write(`${json.logs}`);
         terminalsContext.addTerminal({
-          name: `${resource.namespace.title}: ${container} (logs)`,
+          name: `${resource.namespace}: ${container} (logs)`,
           terminal: term,
         });
         setIsLoading(false);
@@ -154,7 +154,7 @@ const Logs: React.FunctionComponent<ILogsProps> = ({ request, resource, show, se
       if (err.message) {
         term.write(`${err.message}\n\r`);
         terminalsContext.addTerminal({
-          name: `${resource.namespace.title}: ${container} (logs)`,
+          name: `${resource.namespace}: ${container} (logs)`,
           terminal: term,
         });
       }