+++ plugins/resources/src/components/panel/details/actions/Restart.tsx
@@ -1,14 +1,13 @@
 import { AlertVariant, Button, ButtonVariant, Modal, ModalVariant } from '@patternfly/react-core';
-import { IRow } from '@patternfly/react-table';
 import React from 'react';
 import { compare } from 'fast-json-patch';
 
+import { IResource, IResourceRow } from '@kobsio/plugin-core';
 import { IAlert } from '../../../../utils/interfaces';
-import { IResource } from '@kobsio/plugin-core';
 
 interface IRestartProps {
   request: IResource;
-  resource: IRow;
+  resource: IResourceRow;
   show: boolean;
   setShow: (value: boolean) => void;
   setAlert: (alert: IAlert) => void;
@@ -39,9 +38,9 @@ const Restart: React.FunctionComponent<IRestartProps> = ({
       const diff = compare(resource.props, copy);
 
       const response = await fetch(
-        `/api/plugins/resources/resources?cluster=${resource.cluster.title}${
-          resource.namespace ? `&namespace=${resource.namespace.title}` : ''
-        }&name=${resource.name.title}&resource=${request.resource}&path=${request.path}`,
+        `/api/plugins/resources/resources?cluster=${resource.cluster}${
+          resource.namespace ? `&namespace=${resource.namespace}` : ''
+        }&name=${resource.name}&resource=${request.resource}&path=${request.path}`,
         {
           body: JSON.stringify(diff),
           method: 'put',
@@ -51,7 +50,7 @@ const Restart: React.FunctionComponent<IRestartProps> = ({
 
       if (response.status >= 200 && response.status < 300) {
         setShow(false);
-        setAlert({ title: `${resource.name.title} was restarted`, variant: AlertVariant.success });
+        setAlert({ title: `${resource.name} was restarted`, variant: AlertVariant.success });
         refetch();
       } else {
         if (json.error) {
@@ -69,7 +68,7 @@ const Restart: React.FunctionComponent<IRestartProps> = ({
   return (
     <Modal
       variant={ModalVariant.small}
-      title={`Restart ${resource.name.title}`}
+      title={`Restart ${resource.name}`}
       isOpen={show}
       onClose={(): void => setShow(false)}
       actions={[
@@ -82,8 +81,8 @@ const Restart: React.FunctionComponent<IRestartProps> = ({
       ]}
     >
       <p>
-        Do you really want to restart <b>{resource.name.title}</b> (
-        {resource.namespace ? `${resource.namespace.title}/${resource.cluster.title}` : resource.cluster.title})?
+        Do you really want to restart <b>{resource.name}</b> (
+        {resource.namespace ? `${resource.namespace}/${resource.cluster}` : resource.cluster})?
       </p>
     </Modal>
   );